use .net6 project
