+++ PassManager.App/Form1.cs
@@ -1,0 +1,10 @@
+namespace PassManager.App
+{
+    public partial class Form1 : Form
+    {
+        public Form1()
+        {
+            InitializeComponent();
+        }
+    }
+}
