+++ PasswordManager.sln
@@ -1,0 +1,45 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.33627.172
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PassManager.App", "PassManager.App\PassManager.App.csproj", "{0D193225-EFEB-4D8F-BECB-AC56F244DDA9}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{35C6B02F-20DC-4D84-805B-F5E419FB406B}"
+	ProjectSection(SolutionItems) = preProject
+		.gitignore = .gitignore
+		README.md = README.md
+	EndProjectSection
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{1617535B-DBCD-47F9-A137-84A6551D9ED6}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{B8CE02D8-9F02-40DB-971D-6B96B051F6CB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PassManager.App.Test", "test\PassManager.App.Test\PassManager.App.Test.csproj", "{EBE8E8D4-B769-491C-A179-894F3CAF917D}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{0D193225-EFEB-4D8F-BECB-AC56F244DDA9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0D193225-EFEB-4D8F-BECB-AC56F244DDA9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0D193225-EFEB-4D8F-BECB-AC56F244DDA9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0D193225-EFEB-4D8F-BECB-AC56F244DDA9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EBE8E8D4-B769-491C-A179-894F3CAF917D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EBE8E8D4-B769-491C-A179-894F3CAF917D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EBE8E8D4-B769-491C-A179-894F3CAF917D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EBE8E8D4-B769-491C-A179-894F3CAF917D}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{0D193225-EFEB-4D8F-BECB-AC56F244DDA9} = {1617535B-DBCD-47F9-A137-84A6551D9ED6}
+		{EBE8E8D4-B769-491C-A179-894F3CAF917D} = {B8CE02D8-9F02-40DB-971D-6B96B051F6CB}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {00409D55-0F39-42D2-9BBE-2C357CC01DE2}
+	EndGlobalSection
+EndGlobal
