+++ README.md
@@ -1,2 +1,17 @@
 # PasswordManager
 A password management tool that is both free and open source, enabling you to securely manage your passwords.
+
+# Project Structure
+
+**`src`** folder contains the main projects:
+
+- **PaassManager.App** : the main desktop Application (.Net 6.0)
+**src**  folder contains all the main soruce codes /proejcts  
+
+**`test`** folder contains all the test projects
+- **PassManager.App.Test** : xUnit test project (.Net 6.0) of  `PaassManager.App`
+
+
+# Resources
+- [How to Write a Good README File for Your GitHub Project](https://www.freecodecamp.org/news/how-to-write-a-good-readme-file/) 
+- [How to Create Readme file: Best Practices](https://github.com/jehna/readme-best-practices/blob/master/README-default.md)
