--- src/PaasswordManager/PaasswordManager.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.33627.172
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PaassManager.WinApp", "PaassManager.WinApp\PaassManager.WinApp.csproj", "{3021A9DC-5F93-4245-8E3A-B6A56EC061EF}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{DFA1698E-044C-4764-84B9-57DCA97A6C74}"
-	ProjectSection(SolutionItems) = preProject
-		..\..\.gitignore = ..\..\.gitignore
-		..\..\README.md = ..\..\README.md
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3021A9DC-5F93-4245-8E3A-B6A56EC061EF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3021A9DC-5F93-4245-8E3A-B6A56EC061EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3021A9DC-5F93-4245-8E3A-B6A56EC061EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3021A9DC-5F93-4245-8E3A-B6A56EC061EF}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {C22BC4AD-6765-436B-9825-5E329FF95471}
-	EndGlobalSection
-EndGlobal
