+++ test/PassManager.App.Test/UnitTest1.cs
@@ -1,0 +1,11 @@
+namespace PassManager.App.Test
+{
+    public class UnitTest1
+    {
+        [Fact]
+        public void Test1()
+        {
+
+        }
+    }
+}
